integrity_check: initial commit
Add script to check for bitrot using xattrs
Calculates the has of files in the directory you point it at in xattrs
and can be run in a cron job to check a few (bound by runtime) of them every day.

Been running for a couple of years and it hasn't reported anything so
perhaps it has some bugs.